Add Effects module.
