+++ reanimate.cabal
@@ -56,6 +56,7 @@ library
                       Reanimate.Memo
                       Reanimate.Scene
                       Reanimate.Povray
+                      Reanimate.Effect
   other-modules:      Reanimate.Cache
                       Reanimate.Driver.Check
                       Reanimate.Driver.CLI
